chore: move token ts test to src/ts/test
